rename lib to src
