+++ src/WindowManager.js
