+++ src/ipcManager.js
