+++ src/wordCounter.js
